Some OSX build specific changes
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@3529 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ src/corelib/Util.cpp
@@ -238,7 +238,7 @@ size_t _writecbc(FILE *fp, const unsigned char* buffer, int length, unsigned cha
 * and -1 is returned if it matches. (used in V3)
 */
 size_t _readcbc(FILE *fp,
-         unsigned char* &buffer, unsigned int &buffer_len, unsigned char &type,
+         unsigned char* &buffer, size_t &buffer_len, unsigned char &type,
          Fish *Algorithm, unsigned char* cbcbuffer,
          const unsigned char *TERMINAL_BLOCK, size_t file_len)
 {
